+++ Jolt/Physics/Collision/CastSphereVsTriangles.h
@@ -31,7 +31,7 @@ public:
 protected:
 	Vec3							mStart;								///< Starting location of the sphere
 	Vec3							mDirection;							///< Direction and length of movement of sphere
-	float							mRadius;							///< Scaled radius of sphere
+	decimal							mRadius;							///< Scaled radius of sphere
 	const ShapeCastSettings &		mShapeCastSettings;
 	const ShapeFilter &				mShapeFilter;
 	const Mat44 &					mCenterOfMassTransform2;
@@ -40,11 +40,11 @@ protected:
 	CastShapeCollector &			mCollector;
 
 private:
-	void							AddHit(bool inBackFacing, const SubShapeID &inSubShapeID2, float inFraction, Vec3Arg inContactPointA, Vec3Arg inContactPointB, Vec3Arg inContactNormal);
-	void							AddHitWithActiveEdgeDetection(Vec3Arg inV0, Vec3Arg inV1, Vec3Arg inV2, bool inBackFacing, Vec3Arg inTriangleNormal, uint8 inActiveEdges, const SubShapeID &inSubShapeID2, float inFraction, Vec3Arg inContactPointA, Vec3Arg inContactPointB, Vec3Arg inContactNormal);
-	float							RayCylinder(Vec3Arg inRayDirection, Vec3Arg inCylinderA, Vec3Arg inCylinderB, float inRadius) const;
+	void							AddHit(bool inBackFacing, const SubShapeID &inSubShapeID2, decimal inFraction, Vec3Arg inContactPointA, Vec3Arg inContactPointB, Vec3Arg inContactNormal);
+	void							AddHitWithActiveEdgeDetection(Vec3Arg inV0, Vec3Arg inV1, Vec3Arg inV2, bool inBackFacing, Vec3Arg inTriangleNormal, uint8 inActiveEdges, const SubShapeID &inSubShapeID2, decimal inFraction, Vec3Arg inContactPointA, Vec3Arg inContactPointB, Vec3Arg inContactNormal);
+	decimal							RayCylinder(Vec3Arg inRayDirection, Vec3Arg inCylinderA, Vec3Arg inCylinderB, decimal inRadius) const;
 
-	float							mScaleSign;							///< Sign of the scale, -1 if object is inside out, 1 if not
+	decimal							mScaleSign;							///< Sign of the scale, -1 if object is inside out, 1 if not
 };
 
 JPH_NAMESPACE_END
